moved inputs to player class, improved bots going alone logic
