+++ euchre/__main__.py
@@ -6,13 +6,11 @@ from __future__ import annotations
 import time
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from euchre.bots import Bot
     from euchre.cards import Card
     from euchre.dealers import Dealer
     from euchre.players import Player
     from euchre.trumps import Trump
 
-import euchre.bots as _bots
 import euchre.dealers as _dealers
 import euchre.inputs as _inputs
 import euchre.players as _players
@@ -43,11 +41,10 @@ def bidding_round(players: list[Player], dealer: Dealer, revealed: Card=None, pr
         for player in players:
             # delay()
             player.get_player_status()
-            order = _inputs.get_order(revealed)
+            order = player.get_order(revealed)
             if order == 'order' or order == 'yes':
                 trump = _trumps.Trump(revealed.get_suit(), player.get_team())
                 if player.is_bot():
-                    convert_bot_trumps(player, trump)
                     player.going_alone(trump)
                 else:
                     player.going_alone()
@@ -63,21 +60,20 @@ def bidding_round(players: list[Player], dealer: Dealer, revealed: Card=None, pr
             print(f'The dealer {dealer} turned the {previous} face-down. Starting second round of bidding...')
             for player in players:
                 player.get_player_status()
-                call = _inputs.get_call(previous)
+                call = player.get_call(previous)
                 if call == 'pass':
                     continue
                 if call:
-                    player.going_alone()
                     trump = _trumps.Trump(call, player.get_team())
+                    if player.is_bot():
+                        player.going_alone(trump)
+                    else:
+                        player.going_alone()
                     return trump
         else:
             print("ERROR - NO PREVIOUS CARD REFERENCED.")
     return None
 
-def convert_bot_trumps(bot: Bot, trump: Trump):
-    """Converts all bot cards matching trump to the proper value."""
-    for card in bot.get_cards():
-        card.is_trump()
 
 def play_cards(players: list[Player], trump: Trump) -> list[tuple[Player, Card]]:
     """Each player plays a card from their hand. Returns tuple list of (player, card played).
